Include duecredit stub directly
